safedrive-backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6

# Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
# mediapipe==0.10.8
dlib==19.24.2
imutils==0.5.4

# Data Processing
numpy==1.24.3
scipy==1.11.4
pandas==2.1.3

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database (Optional)
# sqlalchemy==2.0.23
# alembic==1.13.0

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2